+++ ft_strnstr.c
@@ -8,12 +8,12 @@ char    *ft_strnstr(const char *s1, const char *s2, size_t n)
     i = 0;
     if (s2[0] == '\0')
         return ((char *)s1);
-    while (s1[i] && i < n)
+    while (s1[i] && i < (int)n)
     {
         j = 0;
         while (s1[i] == s2[j])
         {
-            if (s2[j + 1] == '\0' && i < n)
+            if (s2[j + 1] == '\0' && i < (int)n)
                 return ((char *)&s1[i - j]);
             j++;
             i++;
